Finished story manager and added comments for current classes

+++ MyApplication/app/src/main/java/com/actual/myapplication/IdGenerator.java
@@ -1,14 +1,21 @@
 package com.actual.myapplication;
 
+/**
+ * Manages id variable and the generation process
+ * to ensure a unique id for story objects.
+ *
+ * @author jessiewilkins
+ * @since 2020-02
+ */
 public class IdGenerator {
 
-    private static int id=0;
+    private static long id=0;
 
     private IdGenerator() {
 
     }
 
-    public static int generateId() {
+    public static long generateId() {
         return ++id;
     }
